+++ client/components/date-range-picker/helpers.js
@@ -1,4 +1,5 @@
 import moment from 'moment';
+import { RANGE_OPTIONS } from './constants';
 
 export const getRange = (dateRange) => {
   if (!dateRange) {
@@ -24,6 +25,30 @@ export const getRange = (dateRange) => {
     .toDate();
 
   return [startTime, endTime];
+};
+
+export const getShortcuts = (maxDays, onClickHandler) => {
+  let options = RANGE_OPTIONS.slice();
+
+  if (maxDays && maxDays < 90) {
+    options = options.filter(o => o.daysAgo < maxDays);
+
+    const option = {
+      daysAgo: maxDays,
+      text: `Last ${maxDays} days`,
+      value: `last-${maxDays}-days`,
+    };
+
+    options.push(option);
+    options.sort((a, b) => a.daysAgo - b.daysAgo);
+  }
+
+  options = options.map(option => ({
+    ...option,
+    onClick: () => onClickHandler(option),
+  }));
+
+  return options;
 };
 
 export const isDayDisabled = (maxStartDate) => (date) => {
